==> 06-ASHRAE-great-energy-predictor-III/notebooks/0.1-data-aggregation.ipynb <==
import os
import numpy as np
import pandas as pd

root_dir = os.path.dirname(os.path.abspath(os.getcwd()))         # root directory

# relative paths
bm_path = os.path.join(root_dir, "data", "building_metadata.csv")
train_path = os.path.join(root_dir, "data", "train.csv")
test_path = os.path.join(root_dir, "data", "test.csv")
weather_train_path = os.path.join(root_dir, "data", "weather_train.csv")
weather_test_path = os.path.join(root_dir, "data", "weather_test.csv")

# importing the datasets
bm_df = pd.read_csv(bm_path)
train_df = pd.read_csv(train_path)
test_df = pd.read_csv(test_path)
weather_train_df = pd.read_csv(weather_train_path)
weather_test_df = pd.read_csv(weather_test_path)

bm_df.head()

train_df.head()

test_df.head()

weather_train_df.head()

weather_test_df.head()

train_df = train_df.merge(bm_df, on='building_id', how='left')
train_df = train_df.merge(weather_train_df, on=['site_id', 'timestamp'], how='left')
print("shape of train_df = {}".format(train_df.shape))

test_df = test_df.merge(bm_df, on='building_id', how='left')
test_df = test_df.merge(weather_test_df, on=['site_id', 'timestamp'], how='left')
print("shape of test_df = {}".format(test_df.shape))

train_df.head()

test_df.head()

train_df.to_csv(os.path.join(root_dir, "data", "interim_data", "train_agg.csv"), index = False)
test_df.to_csv(os.path.join(root_dir, "data", "interim_data", "test_agg.csv"), index = False)

def reduce_mem_usage(df, verbose=True):
    numerics = ['int16', 'int32', 'int64', 'float16', 'float32', 'float64']
    start_mem = df.memory_usage().sum() / 1024**2    
    for col in df.columns:
        col_type = df[col].dtypes
        if col_type in numerics:
            c_min = df[col].min()
            c_max = df[col].max()
            if str(col_type)[:3] == 'int':
                if c_min > np.iinfo(np.int8).min and c_max < np.iinfo(np.int8).max:
                    df[col] = df[col].astype(np.int8)
                elif c_min > np.iinfo(np.int16).min and c_max < np.iinfo(np.int16).max:
                    df[col] = df[col].astype(np.int16)
                elif c_min > np.iinfo(np.int32).min and c_max < np.iinfo(np.int32).max:
                    df[col] = df[col].astype(np.int32)
                elif c_min > np.iinfo(np.int64).min and c_max < np.iinfo(np.int64).max:
                    df[col] = df[col].astype(np.int64)  
            else:
                if c_min > np.finfo(np.float16).min and c_max < np.finfo(np.float16).max:
                    df[col] = df[col].astype(np.float16)
                elif c_min > np.finfo(np.float32).min and c_max < np.finfo(np.float32).max:
                    df[col] = df[col].astype(np.float32)
                else:
                    df[col] = df[col].astype(np.float64)    
    end_mem = df.memory_usage().sum() / 1024**2
    if verbose: print('Mem. usage decreased to {:5.2f} Mb ({:.1f}% reduction)'.format(end_mem, 100 * (start_mem - end_mem) / start_mem))
    return df

train_df = reduce_mem_usage(train_df, verbose=True)
test_df = reduce_mem_usage(test_df, verbose=True)

train_df.to_csv(os.path.join(root_dir, "data", "interim_data", "train_agg.csv"), index = False)
test_df.to_csv(os.path.join(root_dir, "data", "interim_data", "test_agg.csv"), index = False)
