==> 06-ASHRAE-great-energy-predictor-III/notebooks/2.1-ml-iteration-xgb.ipynb <==
import os
import json
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns
import xgboost
from sklearn.model_selection import train_test_split
from sklearn.model_selection import KFold
from sklearn.metrics import mean_squared_log_error
from sklearn.metrics import mean_squared_error
from sklearn.metrics import make_scorer
from hyperopt import fmin, hp, tpe, Trials, space_eval, STATUS_OK, STATUS_RUNNING

root_dir = os.path.dirname(os.path.abspath(os.getcwd()))
pd.set_option("display.max_rows", 10, "display.max_columns", None)
plt.rcParams["figure.figsize"] = (20, 12)
sns.set_style("darkgrid")

## Helper Functions

def sr_encoder(dataframe, colname):
    """
    encodes the label using Supervised ratio technique.

    inputs: dataframe and column_name
    output: dataframe with encoded predictor
    """

    label_list = list(dataframe[colname].unique())       # list of all the labels in the predictors
    label_list.sort()                                    # sorting the list

    supervised_ratio = []                                # initializing list and datatype
    label_dict = {}

    for i in label_list:
        event = len(dataframe[dataframe[colname] == i])  # filtering out the colname
        total = len(dataframe)
        SR = round((event/total)*1e3, 3)                 # supervised_ratio
        label_dict[i] = SR                               # forming the dictionary

    dataframe[colname] = dataframe[colname].map(label_dict)

    return dataframe, label_dict

def save_json(dictionary, json_path):
    """
    saves the dictionary for future use. Possibly to encode test values

    input: dict, json file name with extension
    returns nothing
    """

    json_obj = {}
    for k, v in dictionary.items():
        key = str(k)
        json_obj[key] = v

    with open(json_path, "w") as json_file:
        json.dump(json_obj, json_file)

def datetime_engg(dataframe):
    """
    creates new features from datetime column
    """
    dataframe["hour"] = dataframe["timestamp"].dt.hour
    dataframe["day_of_week"] = dataframe["timestamp"].dt.dayofweek
    dataframe["month"] = dataframe["timestamp"].dt.month
    
    return dataframe

train_path = os.path.join(root_dir, "data", "interim_data", "train_agg.csv")
train_df = pd.read_csv(train_path)
train_df.head(5)

# 2. Data Preparation
### 2.1. converting the datatype and filling in Nulls

# converting timestamp to datetime format
train_df["timestamp"] = pd.to_datetime(train_df["timestamp"])

# filling in nulls with -999
train_df = train_df.fillna(-999)

train_df.info()

### 2.2. Encoding the categorical variables
In the dataset, following are the variables that are categorical in nature:-
- `building_id`
- `site_id`
- `primary_use`
- `year_built`

These variables are to be encoded with certain techniques so that we can correctly input the information to the model, reducing the noise.

train_df, puse_dict = sr_encoder(train_df, colname = "primary_use")          # encoding the primary using supervised ratio
train_df, site_dict = sr_encoder(train_df, colname = "site_id")              # encoding the site_id using supervised ratio

# saving the json
save_json(puse_dict, os.path.join(root_dir, "models", "primary_use.json"))
save_json(site_dict, os.path.join(root_dir, "models", "site_id.json"))

train_df.head()

### 2.3. Engineering the Datatime feature

train_df = datetime_engg(train_df)
train_df = train_df.drop(columns = ["timestamp"])

train_df.head()

# 3. Machine Learning Models

## 3.1. XgBoost

x = train_df.drop(columns = ["meter_reading"])
y = train_df["meter_reading"]

del train_df

x_train, x_test, y_train, y_test = train_test_split(x, y, test_size = 0.2, shuffle = False)

del x, y

# define RMSLE
def RMSLE(y_pred, y_true):
    """
    custom eval metric for XGBClassifier
    """
    err = np.sqrt(mean_squared_log_error(y_true, y_pred))
    return "RMSLE", err

def hyperparameter_tuning(params):
    """
    hypertunes a XGBoost model
    
    inp: parameters
    outp: score per fold
    """
    params = {
        "max_depth": int(params["max_depth"]),                              # max depth of the tree
        "gamma": "{:.3f}".format(params["gamma"]),                          # min loss reduction required to make a split
        "subsample": "{:.2f}".format(params["subsample"]),                  # denotes the fraction of observations to be randomly samples for each tree
        "reg_alpha": "{:.3f}".format(params["reg_alpha"]),                  # L1 regularization weight
        "reg_lambda": "{:.3f}".format(params["reg_lambda"]),                # L2 regularization weight
        "learning_rate": "{:.3f}".format(params["learning_rate"]),          # learning rate of XGB
        "num_leaves": "{:.3f}".format(params["num_leaves"]),                 
        "colsample_bytree": "{:.3f}".format(params["colsample_bytree"]),
        "min_child_samples": "{:.3f}".format(params["min_child_samples"]),
        "feature_fraction": "{:.3f}".format(params["feature_fraction"]),
        "bagging_fraction": "{:.3f}".format(params["bagging_fraction"]),
        "objective": "reg:squarederror",
        "RMSLE": RMSLE,
        #"eval_metric": "error"
    }
    
    print("#"*25)
    print("Params = {}".format(params))
    FOLDS = 10  # defining the folds required
    count = 1   # count of HPT cycles
    kf = KFold(n_splits = FOLDS, shuffle = False, random_state = 42)
    y_oof = np.zeros(x_train.shape[0])
    MSLE_mean = 0
    for trn_idx, val_idx in kf.split(x_train):
        # for purpose of saving memory, reducing the size of lists
        trn_idx = list(filter(lambda x: (x % 10 == 0), trn_idx))
        val_idx = list(filter(lambda x: (x % 10 == 0), val_idx))
        
        # define the classifier
        regressor = xgboost.XGBRegressor(random_state = 42, 
                                   verbose = True, 
                                   # tree_method = "gpu_hist",
                                   **params)
        # spliting data into train and valid sets
        train_x, valid_x = x_train.iloc[trn_idx], x_train.iloc[val_idx]
        train_y, valid_y = y_train.iloc[trn_idx], y_train.iloc[val_idx]
        
        # fit the estimator and predict
        regressor.fit(train_x, train_y)
        pred = regressor.predict(valid_x)
        pred = [0 for i in pred if i < 0]
        # eval metrics
        score_MSLE = make_scorer(mean_squared_error)(regressor, valid_x, valid_y)
        MSLE_mean += score_MSLE
        print("Count = {} ... score_MSLE = {:.4f}".format(count, score_MSLE))
        count += 1
    
    gc.collect()
    print("Mean MSLE = {:.4}".format(MSLE_mean / FOLDS))
    del train_x, valid_x, train_y, valid_y, regressor, score_MSLE
    
    return -(MLSE_mean/FOLDS)

space = {
    "max_depth": hp.quniform("max_depth", 6, 8, 1),
    "reg_alpha": hp.uniform("reg_alpha", 0.01, 0.05),
    "reg_lambda": hp.uniform("reg_lambda", 0.01, 0.05),
    "learning_rate": hp.uniform("learning_rate", 0.001, 0.2),
    "colsample_bytree": hp.uniform("colsample_bytree", 0.3, 0.9),
    "gamma": hp.uniform("gamma", 0.01, 0.7),
    "num_leaves": hp.choice("num_leaves", list(range(20, 250, 10))),
    "min_child_samples": hp.choice("min_child_samples", list(range(100, 250, 10))),
    "subsample": hp.choice('subsample', [0.2, 0.4, 0.5, 0.6, 0.7, .8, .9]),
    'feature_fraction': hp.uniform('feature_fraction', 0.4, .8),
    'bagging_fraction': hp.uniform('bagging_fraction', 0.4, .9)
}

%%time

# Set algoritm parameters
best = fmin(fn = hyperparameter_tuning,
            space = space,
            algo = tpe.suggest,
            max_evals = 5)

# Print best parameters
best_params = space_eval(space, best)
