==> 06-ASHRAE-great-energy-predictor-III/notebooks/05_Creating_CV_set.ipynb <==
import numpy as np
import pandas as pd
import gc, warnings, random, math
import pickle

from sklearn.metrics import mean_squared_error
from sklearn.model_selection import train_test_split

warnings.filterwarnings('ignore')
pd.set_option('display.max_columns', 35)

SEED = 17
TARGET = 'meter_reading'
random.seed(SEED)
np.random.seed(SEED)

def rmse(y, y_pred):
    return np.sqrt(mean_squared_error(y, y_pred))

df_train = pd.read_pickle('df_train.pkl')
df_test = pd.read_pickle('df_test.pkl')

from sklearn.preprocessing import LabelEncoder

le = LabelEncoder()
df_train['primary_use'] = df_train['primary_use'].astype(str)
df_train['primary_use'] = le.fit_transform(df_train['primary_use']).astype(np.int8)

for df in [df_train, df_test]:
    
    df['DT_D'] = df['timestamp'].dt.dayofyear.astype(np.int16)
    df['DT_M'] = df['timestamp'].dt.month.astype(np.int8)
    df['DT_hour'] = df['timestamp'].dt.hour.astype(np.int8)
    df["weekday"] = df['timestamp'].dt.weekday

df_train.head()

temp = df_test[~df_test['building_id'].isin(df_train['building_id'])]
print(f"{len(temp)}: Buildings are exclusive in test set")
temp = df_test[~df_test['site_id'].isin(df_train['site_id'])]
print(f"{len(temp)}: Sites are exclusive in test set")
del temp, df_test
gc.collect()

## Creating Holdout sets

#1st Holdout set
# Split the train set by site_id -> ~20% in the validation set
train,test = train_test_split(df_train['site_id'].unique(), test_size=0.2, random_state=SEED)
holdout_set1 = df_train[df_train['site_id'].isin(test)].reset_index(drop=True)
df_train = df_train[df_train['site_id'].isin(train)].reset_index(drop=True)

#2nd Holdout set
#Split the train set by building_id -> ~20% in the validation set
train,test = train_test_split(df_train['building_id'].unique(), test_size=0.2, random_state=SEED)
holdout_set2 = df_train[df_train['building_id'].isin(test)].reset_index(drop=True)
df_train = df_train[df_train['building_id'].isin(train)].reset_index(drop=True)

#3rd Holdout set
#Split the train set by month -> First and Last Month in holdout set
holdout_set3 = df_train[(df_train['DT_M']==1) | (df_train['DT_M'] ==12)].reset_index(drop=True)
df_train = df_train[(df_train['DT_M']!=1) & (df_train['DT_M']) !=12].reset_index(drop=True)

gc.collect()

for df in [df_train, holdout_set1, holdout_set2, holdout_set3]:
    df['meter_reading'] = np.log1p(df['meter_reading'])
    print(df.shape)

remove_columns = ['timestamp', 'site_id', 'building_id', TARGET]
features_columns = [col for col in list(df_train) if col not in remove_columns]

X = df_train[features_columns]
y = df_train[TARGET]

split_by_building = df_train['building_id']
split_by_site = df_train['site_id']
split_by_month = df_train['DT_M']

del df_train
gc.collect()

READING_1 = holdout_set1[[TARGET]]
READING_2 = holdout_set2[[TARGET]]
READING_3 = holdout_set3[[TARGET]]

all_readings = {
    1:[READING_1, holdout_set1, 'site_id_hold_out'],
    2:[READING_2, holdout_set2, 'build_id_hold_out'],
    3:[READING_3, holdout_set3, 'month_hold_out']
}

for _,df in all_readings.items():
    df[0]['test'] = 0    
    print('Ground RMSE for', df[2], '|',
          rmse(df[0][TARGET], df[0]['test']))
    del df[0]['test']
    print('#'*50) 

## WE will be using same number of splits for traning the model
N_SPLITS = 5

import lightgbm as lgb

params = {
                    'objective':'regression',
                    'boosting_type':'gbdt',
                    'metric':'rmse',
                    'n_jobs':-1,
                    'learning_rate':0.3, #for faster training
                    'num_leaves': 2**8,
                    'max_depth':-1,
                    'tree_learner':'serial',
                    'colsample_bytree': 0.7,
                    'subsample_freq':1,
                    'subsample':0.7,
                    'n_estimators':1000,
                    'max_bin':255,
                    'verbose':-1,
                    'seed': SEED,
                    'early_stopping_rounds':100, 
                } 

#Trying Multiple bossting round to find the best among them
for n_rounds in [25,50,100,200]:
    
    print('#'*50)
    print('No Validation training...', n_rounds, 'boosting rounds')
    corrected_lgb_params = params.copy()
    corrected_lgb_params['n_estimators'] = n_rounds
    corrected_lgb_params['early_stopping_rounds'] = None

    train_data = lgb.Dataset(X, label=y)
    
    estimator = lgb.train(
                corrected_lgb_params,
                train_data
            )
    gc.collect()
    for _,df in all_readings.items():
        df[0]['no_validation_'+str(n_rounds)] = estimator.predict(df[1][features_columns])
        print('RMSE for',
              df[2], '|',
              rmse(df[0][TARGET], df[0]['no_validation_'+str(n_rounds)]))
        print('#'*20)

__Results suggest there is data leakage by in our dataset__ 
* leakage by Site_ID ->> Model Performs poorly on unknown sites hence doesn't generalize well 
* leakage by Building_IX ->> Model Performs poorly on unknown buildings hence doesn't generalize well 

Luckily our train set has same site and building ids as that of test set. Probably we don't need to smooth differences between them and can even make differences more explicit.
So, a good CV set should have all buildings and site information as bulidings and sites have very unqiue energy consumption.

gc.collect()

print('#'*20)
print('KFold (with shuffle) training...')

from sklearn.model_selection import KFold
folds = KFold(n_splits=N_SPLITS, shuffle=True, random_state=SEED)

for _,df in all_readings.items():
    df[0]['shuffle_kfold'] = 0
        
for fold_, (trn_idx, val_idx) in enumerate(folds.split(X, y)):
    print('Fold:',fold_+1)
    tr_x, tr_y = X.iloc[trn_idx,:], y[trn_idx]    
    vl_x, v_y = X.iloc[val_idx,:], y[val_idx]    
    train_data = lgb.Dataset(tr_x, label=tr_y)
    valid_data = lgb.Dataset(vl_x, label=v_y)  

    estimator = lgb.train(
            params,
            train_data,
            valid_sets = [train_data, valid_data],
            verbose_eval = 100,
        )
    
    for _,df in all_readings.items():
        df[0]['shuffle_kfold'] += estimator.predict(df[1][features_columns])/N_SPLITS

for _,df in all_readings.items():
    print('RMSE for', df[2], '|',
          rmse(df[0][TARGET], df[0]['shuffle_kfold']))
    print('#'*20) 

__Note: Changing the number of splits 5 to 3, as it is taking quite a bit of time, but will use 5-Fold for the traning the final model__

print('#'*20)
print('KFold (no shuffle) training...')

from sklearn.model_selection import KFold
folds = KFold(n_splits=N_SPLITS, shuffle=False, random_state=SEED)

for _,df in all_readings.items():
    df[0]['no_shuffle_kfold'] = 0
        
for fold_, (trn_idx, val_idx) in enumerate(folds.split(X, y)):
    print('Fold:',fold_+1)
    tr_x, tr_y = X.iloc[trn_idx,:], y[trn_idx]    
    vl_x, v_y = X.iloc[val_idx,:], y[val_idx]    
    train_data = lgb.Dataset(tr_x, label=tr_y)
    valid_data = lgb.Dataset(vl_x, label=v_y)  

    estimator = lgb.train(
            params,
            train_data,
            valid_sets = [train_data, valid_data],
            verbose_eval = 100,
        )
    
    for _,df in all_readings.items():
        df[0]['no_shuffle_kfold'] += estimator.predict(df[1][features_columns])/N_SPLITS

for _,df in all_readings.items():
    print('RMSE for', df[2], '|',
          rmse(df[0][TARGET], df[0]['no_shuffle_kfold']))
    print('#'*20) 

__Note: Data Leakage by month: Consumption differ from month to month on a quite significant magnitude, hence we cannot exclude any data by month.__

print('#'*20)
print('GroupKFold site_id split training...') 

from sklearn.model_selection import GroupKFold
folds = GroupKFold(n_splits=N_SPLITS)

for _,df in all_readings.items():
    df[0]['Groupkfold_by_site'] = 0
      
for fold_, (trn_idx, val_idx) in enumerate(folds.split(X, y, groups=split_by_site)):
    print('Fold:',fold_+1)
    tr_x, tr_y = X.iloc[trn_idx,:], y[trn_idx]    
    vl_x, v_y = X.iloc[val_idx,:], y[val_idx]    
    train_data = lgb.Dataset(tr_x, label=tr_y)
    valid_data = lgb.Dataset(vl_x, label=v_y)  

    estimator = lgb.train(
            params,
            train_data,
            valid_sets = [train_data, valid_data],
            verbose_eval = 100,
        )

    for _,df in all_readings.items():
        df[0]['Groupkfold_by_site'] += estimator.predict(df[1][features_columns])/N_SPLITS

for _,df in all_readings.items():
    print('RMSE for', df[2], '|',
          rmse(df[0][TARGET], df[0]['Groupkfold_by_site']))
    print('#'*20)  

print('#'*20)
print('GroupKFold building_id split training...') 

from sklearn.model_selection import GroupKFold
folds = GroupKFold(n_splits=N_SPLITS)

for _,df in all_readings.items():
    df[0]['Groupkfold_by_building'] = 0
      
for fold_, (trn_idx, val_idx) in enumerate(folds.split(X, y, groups=split_by_building)):
    print('Fold:',fold_+1)
    tr_x, tr_y = X.iloc[trn_idx,:], y[trn_idx]    
    vl_x, v_y = X.iloc[val_idx,:], y[val_idx]    
    train_data = lgb.Dataset(tr_x, label=tr_y)
    valid_data = lgb.Dataset(vl_x, label=v_y)  

    estimator = lgb.train(
            params,
            train_data,
            valid_sets = [train_data, valid_data],
            verbose_eval = 100,
        )

    for _,df in all_readings.items():
        df[0]['Groupkfold_by_building'] += estimator.predict(df[1][features_columns])/N_SPLITS

for _,df in all_readings.items():
    print('RMSE for', df[2], '|',
          rmse(df[0][TARGET], df[0]['Groupkfold_by_building']))
    print('#'*20)  

print('#'*20)
print('GroupKFold month split training...') 

from sklearn.model_selection import GroupKFold
folds = GroupKFold(n_splits=N_SPLITS)

for _,df in all_readings.items():
    df[0]['Groupkfold_by_month'] = 0
      
for fold_, (trn_idx, val_idx) in enumerate(folds.split(X, y, groups=split_by_month)):
    print('Fold:',fold_+1)
    tr_x, tr_y = X.iloc[trn_idx,:], y[trn_idx]    
    vl_x, v_y = X.iloc[val_idx,:], y[val_idx]    
    train_data = lgb.Dataset(tr_x, label=tr_y)
    valid_data = lgb.Dataset(vl_x, label=v_y)  

    estimator = lgb.train(
            params,
            train_data,
            valid_sets = [train_data, valid_data],
            verbose_eval = 100,
        )

    for _,df in all_readings.items():
        df[0]['Groupkfold_by_month'] += estimator.predict(df[1][features_columns])/N_SPLITS

for _,df in all_readings.items():
    print('RMSE for', df[2], '|',
          rmse(df[0][TARGET], df[0]['Groupkfold_by_month']))
    print('#'*20)

### Findings:

* Dataleakage is preventing model to generalize well.
* The best way to divide the data will be on the month basis, using 10 months for traning and 2 months for validation
* Timesatamp has to be fixed
* More boosting rounds with early stopping is to be used
* Have to train several models on different SEEDS and have to average them out

Train set - first 4 month
Skip - next 4 month
Valid set - last 4 month
