==> 06-ASHRAE-great-energy-predictor-III/notebooks/06_meter_ensemble_50_50.ipynb <==
import numpy as np
import pandas as pd
import warnings
import gc

import matplotlib.pyplot as plt
import seaborn as sns
%matplotlib inline
color = sns.color_palette()
sns.set_style("darkgrid")
warnings.filterwarnings('ignore')

df_train = pd.read_pickle('df_train.pkl')

df_train.sample(5)

print(f"Max Air Temp:{np.max(df_train['air_temperature'])}")
print(f"Min Air Temp:{np.min(df_train['air_temperature'])}")
print(f"Max Dew Temp:{np.max(df_train['dew_temperature'])}")
print(f"Min Dew Temp:{np.min(df_train['dew_temperature'])}")
print(f"Max Precipitation Depth:{np.max(df_train['precip_depth_1_hr'])}")
print(f"Min Precipitation Depth:{np.min(df_train['precip_depth_1_hr'])}")

df_train.dtypes

df_train.isna().sum()

from sklearn.preprocessing import LabelEncoder

le = LabelEncoder()
df_train['primary_use'] = le.fit_transform(df_train['primary_use'])

df_train_0 = df_train[df_train['meter']==0]
df_train_1 = df_train[df_train['meter']==1]
df_train_2 = df_train[df_train['meter']==2]
df_train_3 = df_train[df_train['meter']==3]

del df_train
gc.collect()

# print(f"No of Unique Primary use for meter{0} is: {df_train_0['primary_use'].nunique()}")
# print(f"No of Unique Primary use for meter{1} is: {df_train_1['primary_use'].nunique()}")
# print(f"No of Unique Primary use for meter{2} is: {df_train_2['primary_use'].nunique()}")
# print(f"No of Unique Primary use for meter{3} is: {df_train_3['primary_use'].nunique()}")

def prepare_data(X, test=False):
    """
    Preparing final dataset with all features.
    """
    X.square_feet = np.log1p(X.square_feet)
    
    if not test:
        X.sort_values("timestamp", inplace=True)
        X.reset_index(drop=True, inplace=True)
    
    gc.collect()
    
    holidays = ["2016-01-01", "2016-01-18", "2016-02-15", "2016-05-30", "2016-07-04",
                "2016-09-05", "2016-10-10", "2016-11-11", "2016-11-24", "2016-12-26",
                "2017-01-01", "2017-01-16", "2017-02-20", "2017-05-29", "2017-07-04",
                "2017-09-04", "2017-10-09", "2017-11-10", "2017-11-23", "2017-12-25",
                "2018-01-01", "2018-01-15", "2018-02-19", "2018-05-28", "2018-07-04",
                "2018-09-03", "2018-10-08", "2018-11-12", "2018-11-22", "2018-12-25",
                "2019-01-01"]
    
    X["hour"] = X.timestamp.dt.hour
    X["weekday"] = X.timestamp.dt.weekday
    X["is_holiday"] = (X.timestamp.dt.date.astype("str").isin(holidays)).astype(int)
    
    drop_features = ["timestamp", "sea_level_pressure", "wind_direction", "wind_speed"]
    X.drop(drop_features, axis=1, inplace=True)
    
    ####imputing NaNs
    X['year_built'].fillna(-1, inplace=True)
    X['floor_count'].fillna(0, inplace=True)
    X['air_temperature'].fillna(0, inplace=True)
    X['cloud_coverage'].fillna(-1, inplace=True)
    X['dew_temperature'].fillna(0, inplace=True)
    X['precip_depth_1_hr'].fillna(-2, inplace=True)
    
    if test:
        return X
#         row_ids = X.row_id
#         X.drop("row_id", axis=1, inplace=True)
#         return X, row_ids
    else:
        y = np.log1p(X.meter_reading)
        X.drop("meter_reading", axis=1, inplace=True)
        return X, y

X_0, y_0 = prepare_data(df_train_0)
X_1, y_1 = prepare_data(df_train_1)
X_2, y_2 = prepare_data(df_train_2)
X_3, y_3 = prepare_data(df_train_3)

del df_train_0, df_train_1, df_train_2, df_train_3
gc.collect()

X_0.head(10)

X_0_1 =  X_0[:int(X_0.shape[0] / 2)]
X_0_2 =  X_0[int(X_0.shape[0] / 2):]
X_1_1 =  X_0[:int(X_1.shape[0] / 2)]
X_1_2 =  X_0[int(X_1.shape[0] / 2):]
X_2_1 =  X_0[:int(X_2.shape[0] / 2)]
X_2_2 =  X_0[int(X_2.shape[0] / 2):]
X_3_1 =  X_0[:int(X_3.shape[0] / 2)]
X_3_2 =  X_0[int(X_3.shape[0] / 2):]

y_0_1 = y_0[:int(X_0.shape[0] / 2)]
y_0_2 = y_0[int(X_0.shape[0] / 2):]
y_1_1 = y_0[:int(X_1.shape[0] / 2)]
y_1_2 = y_0[int(X_1.shape[0] / 2):]
y_2_1 = y_0[:int(X_2.shape[0] / 2)]
y_2_2 = y_0[int(X_2.shape[0] / 2):]
y_3_1 = y_0[:int(X_3.shape[0] / 2)]
y_3_2 = y_0[int(X_3.shape[0] / 2):]

del X_0, y_0, X_1, y_1, X_2, y_2, X_3, y_3
gc.collect()

import lightgbm as lgb
import random
from sklearn.metrics import mean_squared_error
random.seed(12)

categorical_features = ["building_id", "site_id", "meter", "primary_use", "hour", "weekday"]

d_0_1 = lgb.Dataset(X_0_1, label=y_0_1, categorical_feature=categorical_features, free_raw_data=False)
d_0_2 = lgb.Dataset(X_0_2, label=y_0_2, categorical_feature=categorical_features, free_raw_data=False)
d_1_1 = lgb.Dataset(X_1_1, label=y_1_1, categorical_feature=categorical_features, free_raw_data=False)
d_1_2 = lgb.Dataset(X_1_2, label=y_1_2, categorical_feature=categorical_features, free_raw_data=False)
d_2_1 = lgb.Dataset(X_2_1, label=y_2_1, categorical_feature=categorical_features, free_raw_data=False)
d_2_2 = lgb.Dataset(X_2_2, label=y_2_2, categorical_feature=categorical_features, free_raw_data=False)
d_3_1 = lgb.Dataset(X_3_1, label=y_3_1, categorical_feature=categorical_features, free_raw_data=False)
d_3_2 = lgb.Dataset(X_3_2, label=y_3_2, categorical_feature=categorical_features, free_raw_data=False)

watchlist_0_1 = [d_0_1, d_0_2]
watchlist_0_2 = [d_0_2, d_0_1]
watchlist_1_1 = [d_1_1, d_1_2]
watchlist_1_2 = [d_1_2, d_1_1]
watchlist_2_1 = [d_2_1, d_2_2]
watchlist_2_2 = [d_2_2, d_2_1]
watchlist_3_1 = [d_3_1, d_3_2]
watchlist_3_2 = [d_3_2, d_3_1]

params = {
    "objective": "regression",
    "boosting": "gbdt",
    "num_leaves": 40,
    "learning_rate": 0.05,
    "feature_fraction": 0.85,
    "reg_lambda": 2,
    "metric": "rmse"
}

print("Building LGBM Model with for meter 0 with 1st half of data and validation on 2nd half:")
model_0_1 = lgb.train(params, train_set=d_0_1, 
                         num_boost_round=1000, valid_sets=watchlist_0_1, 
                         verbose_eval=200, early_stopping_rounds=200)
print('#'*50) 

print("Building LGBM Model with for meter 0 with 2nd half of data and validation on 1st half:")
model_0_2 = lgb.train(params, train_set=d_0_2, 
                         num_boost_round=1000, valid_sets=watchlist_0_2, 
                         verbose_eval=200, early_stopping_rounds=200)
print('#'*50) 

print("Building LGBM Model with for meter 1 with 1st half of data and validation on 2nd half:")
model_1_1 = lgb.train(params, train_set=d_1_1, 
                         num_boost_round=1000, valid_sets=watchlist_1_1, 
                         verbose_eval=200, early_stopping_rounds=200)
print('#'*50) 

print("Building LGBM Model with for meter 1 with 2nd half of data and validation on 1st half:")
model_1_2 = lgb.train(params, train_set=d_1_2, 
                         num_boost_round=1000, valid_sets=watchlist_1_2, 
                         verbose_eval=200, early_stopping_rounds=200)
print('#'*50) 

print("Building LGBM Model with for meter 2 with 1st half of data and validation on 2nd half:")
model_2_1 = lgb.train(params, train_set=d_2_1, 
                         num_boost_round=1000, valid_sets=watchlist_2_1, 
                         verbose_eval=200, early_stopping_rounds=200)
print('#'*50) 

print("Building LGBM Model with for meter 2 with 2nd half of data and validation on 1st half:")
model_2_2 = lgb.train(params, train_set=d_2_2, 
                         num_boost_round=1000, valid_sets=watchlist_2_2, 
                         verbose_eval=200, early_stopping_rounds=200)
print('#'*50) 

print("Building LGBM Model with for meter 3 with 1st half of data and validation on 2nd half:")
model_3_1 = lgb.train(params, train_set=d_3_1, 
                         num_boost_round=1000, valid_sets=watchlist_3_1, 
                         verbose_eval=200, early_stopping_rounds=200)
print('#'*50) 

print("Building LGBM Model with for meter 3 with 2nd half of data and validation on 1st half:")
model_3_2 = lgb.train(params, train_set=d_3_2, 
                         num_boost_round=1000, valid_sets=watchlist_3_2, 
                         verbose_eval=200, early_stopping_rounds=200)

models = {'model_0_1' : model_0_1,
          'model_0_2' : model_0_2,
          'model_1_1' : model_1_1,
          'model_1_2' : model_1_2,
          'model_2_1' : model_2_1,
          'model_2_2' : model_2_2,
          'model_3_1' : model_3_1,
          'model_3_2' : model_3_2 }

import pickle

for i in range(4):
    for j in range(1,3):
        picfile = open(f'model_{i}_{j}', 'wb') 
        pickle.dump(models[f'model_{i}_{j}'], picfile)
        picfile.close() 

picfile = open('le', 'wb') 
pickle.dump(le, picfile)
picfile.close() 

del X_0_1, X_0_2, X_1_1, X_1_2, X_2_1, X_2_2, X_3_1, X_3_2
del y_0_1, y_0_2, y_1_1, y_1_2, y_2_1, y_2_2, y_3_1, y_3_2
    
gc.collect()

df_test = pd.read_pickle('df_test.pkl')

models = {}
for i in range(4):
    for j in range(1,3):
        picfile = open(f'model_{i}_{j}', 'rb') 
        models[f'model_{i}_{j}'] = pickle.load(picfile)
        picfile.close() 
        
picfile = open('le', 'rb') 
le = pickle.load(picfile)
picfile.close() 

df_test.primary_use = le.transform(df_test.primary_use)

# X_test, row_ids = prepare_data(df_test, test=True)
X_test = prepare_data(df_test, test=True)
del df_test
gc.collect()

for i in range(4):
    temp = X_test[X_test['meter'] == i]
    row = temp['row_id']
    temp.drop('row_id', axis=1, inplace=True)
    for j in range(1,3):
        if j == 1:
            pred = np.expm1(models[f'model_{i}_{j}'].predict(temp, num_iteration=models[f'model_{i}_{j}'].best_iteration)) / 2
            del models[f'model_{i}_{j}']
            gc.collect()
        else:
            pred += np.expm1(models[f'model_{i}_{j}'].predict(temp, num_iteration=models[f'model_{i}_{j}'].best_iteration)) / 2
            del models[f'model_{i}_{j}']
            gc.collect()
    
    pd.DataFrame({"row_id": row, "meter_reading": np.clip(pred, 0, a_max=None)}).to_csv(f"submission{i}.csv", index=False)
    del temp,  row, pred
    gc.collect()

del X_test
gc.collect()

sub0 = pd.read_csv('submission0.csv')
sub1 = pd.read_csv('submission1.csv')
sub2 = pd.read_csv('submission2.csv')
sub3 = pd.read_csv('submission3.csv')

frames = [sub0, sub1, sub2, sub3]

result = pd.concat(frames)
del sub0, sub1, sub2, sub3
gc.collect()

result.shape

result.to_csv('submission.csv', index=False)
